requirements.txt
fastapi==0.95.0
uvicorn==0.22.0
streamlit==1.25.0
openai==0.27.0
requests==2.31.0
python-dotenv==0.19.0
